requirements.txt
django>=1.6
django-braces>=1.4.0
django-timezone-field>=1.0
south>=0.8.4

# Tests Require
mock
django-dynamic-fixture

